fix some gulp wiring
