+++ packages/scripts/src/commands/_lint/tasks.ts
@@ -2,37 +2,21 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import { performance } from 'perf_hooks'
 import gulp from 'gulp'
 import { LintCommandOptions } from './types'
-import { eslint } from '@essex/build-step-eslint'
-import { prettyQuick } from '@essex/build-step-pretty-quick'
-import { resolveGulpTask } from '@essex/build-utils'
+import { eslintGulp as eslint } from '@essex/build-step-eslint'
+import { prettyQuickGulp as prettyQuick } from '@essex/build-step-pretty-quick'
+import { gulpify, wrapPromiseTask } from '@essex/build-utils'
 
 export function configureTasks(
-	{
-		fix = false,
-		staged = false,
-		docs = false,
-		strict = false,
-		docsOnly = false,
-	}: LintCommandOptions,
+	{ fix = false, staged = false, strict = false }: LintCommandOptions,
 	files: string[] | undefined,
 ): gulp.TaskFunction {
-	function checkCode(cb: (err?: Error) => void) {
-		const start = performance.now()
-		eslint(fix, strict, files || ['.']).then(
-			...resolveGulpTask('eslint', start, cb),
-		)
-	}
+	const checkCode = eslint(fix, strict, files || ['.'])
 
-	function checkFormatting(cb: (err?: Error) => void) {
-		const start = performance.now()
-		const task = staged
-			? prettyQuick({ staged: true })
-			: prettyQuick({ check: !fix })
-		task.then(...resolveGulpTask('pretty-quick', start, cb))
-	}
+	const checkFormatting = staged
+		? prettyQuick({ staged: true })
+		: prettyQuick({ check: !fix })
 
 	return gulp.parallel(checkCode, checkFormatting)
 }
