+++ packages/step-babel/package.json
@@ -15,6 +15,7 @@
 		"@babel/core": "^7.14.3",
 		"@babel/types": "^7.14.4",
 		"@essex/babel-config": "workspace:packages/config-babel",
+		"@essex/build-utils": "workspace:packages/util-utils",
 		"@essex/tasklogger": "workspace:packages/util-tasklogger",
 		"chalk": "^4.1.1",
 		"glob": "^7.1.7",
