+++ packages/step-clean/src/index.ts
@@ -4,11 +4,11 @@
  */
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import { rm } from './rm'
-import { gulpify } from '@essex/build-utils'
+import { gulpify, wrapPromiseTask } from '@essex/build-utils'
 
 export async function clean(files: string[]): Promise<void> {
 	const promises: Array<Promise<void>> = files.filter(f => !!f).map(f => rm(f!))
 	await Promise.all(promises)
 }
 
-export const cleanGulp = gulpify('clean', clean)
+export const cleanGulp = gulpify(wrapPromiseTask('clean', false, clean))
