+++ packages/step-eslint/src/index.ts
@@ -5,7 +5,7 @@
 /* eslint-disable @essex/adjacent-await */
 import { ESLint } from 'eslint'
 import { getConfigFile, getIgnorePath } from './config'
-import { gulpify } from '@essex/build-utils'
+import { gulpify, wrapPromiseTask } from '@essex/build-utils'
 
 export async function eslint(
 	fix: boolean,
@@ -47,4 +47,4 @@ export async function eslint(
 	}
 }
 
-export const eslintGulp = gulpify('eslint', eslint)
+export const eslintGulp = gulpify(wrapPromiseTask('eslint', false, eslint))
