+++ packages/step-pretty-quick/src/index.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 import chalk from 'chalk'
-import { gulpify } from '@essex/build-utils'
+import { gulpify, wrapPromiseTask } from '@essex/build-utils'
 
 /* eslint-disable-next-line @typescript-eslint/no-var-requires */
 const pq = require('pretty-quick').default
@@ -78,4 +78,6 @@ export function prettyQuick(args: PrettyQuickArgs): Promise<void> {
 	}
 }
 
-export const prettyQuickGulp = gulpify('pretty-quick', prettyQuick)
+export const prettyQuickGulp = gulpify(
+	wrapPromiseTask('pretty-quick', false, prettyQuick),
+)
