+++ packages/util-utils/package.json
@@ -21,7 +21,7 @@
 	},
 	"devDependencies": {
 		"@tsconfig/node12": "^1.0.7",
-		"@types/node": "^14.17.2",
+		"@types/node": "14",
 		"rimraf": "^3.0.2",
 		"typescript": "^4.3.2"
 	}
