+++ packages/util-utils/src/index.ts
@@ -3,10 +3,10 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 export * from './noop'
-export * from './resolveGulpTask'
 export * from './filterShellCode'
 export * from './execGulpTask'
 export * from './types'
-export * from './gulpify'
 export * from './streamToPromise'
 export * from './resolveTask'
+export * from './wrapPromiseTask'
+export * from './gulpify'
