+++ packages/util-utils/src/noop.ts
@@ -3,5 +3,6 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 import through2 from 'through2'
+import { GulpyTask } from './types'
 export const noopStep = (): any => through2.obj()
-export const noopTask = (cb: (err?: Error) => void): void => cb()
+export const noopTask: GulpyTask = (cb: (err?: Error) => void): void => cb()
